백앤드 API 오류 수정 및 프론트 연결 로그인 기능 추가 작업

+++ backend/src/main/java/com/gameproject/backend/dto/DialogueReplyResponse.java
@@ -3,6 +3,6 @@ package com.gameproject.backend.dto;
 public record DialogueReplyResponse(
         String npcReply,
         Integer affinityScore,
-        Integer staminaCurrent
+        Double staminaCurrent
 ) {
 }
