+++ backend/src/main/java/com/gameproject/backend/dto/ForageResponse.java
@@ -3,6 +3,6 @@ package com.gameproject.backend.dto;
 public record ForageResponse(
         Long fruitId,
         String fruitName,
-        Integer staminaCurrent
+        Double staminaCurrent
 ) {
 }
